# Use session and combine lock with door status