--- a/august/api.py
+++ b/august/api.py
@@ -82,9 +82,27 @@ def _determine_lock_door_status(status):
 
 
 class Api:
-    def __init__(self, timeout=10, command_timeout=60):
+    def __init__(self, timeout=10, command_timeout=60, use_http_session=False):
         self._timeout = timeout
         self._command_timeout = command_timeout
+        self._http_session = requests.Session() if use_http_session else None
+
+    def __del__(self):
+        """ Close the session if exist when this instance is destroyed
+
+        Not liked by all or everyone, but implementing this as an "in-case"
+        When Api class is called with use_http_session=True then method close
+        should be called to close the session before removing last reference
+        """
+        try:
+            self.close()
+        except Exception:
+            pass
+
+    def close(self):
+        """ Close the session """
+        if self._http_session is not None:
+            self._http_session.close()
 
     def get_session(self, install_id, identifier, password):
         response = self._call_api(
@@ -207,21 +225,29 @@ class Api:
 
         return LockDetail(response.json())
 
-    def get_lock_status(self, access_token, lock_id):
+    def get_lock_status(self, access_token, lock_id, door_status=False):
         json = self._call_api(
             "get",
             API_GET_LOCK_STATUS_URL.format(lock_id=lock_id),
             access_token=access_token).json()
 
-        return _determine_lock_status(json["status"])
+        if door_status:
+            return (_determine_lock_status(json.get("status")),
+                    _determine_lock_door_status(json.get("doorState")))
 
-    def get_lock_door_status(self, access_token, lock_id):
+        return _determine_lock_status(json.get("status"))
+
+    def get_lock_door_status(self, access_token, lock_id, lock_status=False):
         json = self._call_api(
             "get",
             API_GET_LOCK_STATUS_URL.format(lock_id=lock_id),
             access_token=access_token).json()
 
-        return _determine_lock_door_status(json["doorState"])
+        if lock_status:
+            return (_determine_lock_door_status(json.get("doorState")),
+                    _determine_lock_status(json.get("status")))
+
+        return _determine_lock_door_status(json.get("doorState"))
 
     def lock(self, access_token, lock_id):
         json = self._call_api(
@@ -230,7 +256,7 @@ class Api:
             access_token=access_token,
             timeout=self._command_timeout).json()
 
-        return _determine_lock_status(json["status"])
+        return _determine_lock_status(json.get("status"))
 
     def unlock(self, access_token, lock_id):
         json = self._call_api(
@@ -239,7 +265,7 @@ class Api:
             access_token=access_token,
             timeout=self._command_timeout).json()
 
-        return _determine_lock_status(json["status"])
+        return _determine_lock_status(json.get("status"))
 
     def _call_api(self, method, url, access_token=None, **kwargs):
         payload = kwargs.get("params") or kwargs.get("json")
@@ -253,7 +279,9 @@ class Api:
         _LOGGER.debug("About to call %s with header=%s and payload=%s", url,
                       kwargs["headers"], payload)
 
-        response = requests.request(method, url, **kwargs)
+        response = self._http_session.request(method, url, **kwargs) if\
+            self._http_session is not None else\
+            requests.request(method, url, **kwargs)
 
         _LOGGER.debug("Received API response: %s, %s", response.status_code,
                       response.content)