@snjoetw 

This change will use session from Requests to re-use the same connection when retrieving status (or sending commands). Also made a change to allow one to retrieve both lock and door state in 1 call (optional) as for both the information is retrieved from the same URL in August. 

Did testing to determine what the performance difference would be:

```
Test:
	100 iterations to retrieve door_status, lock_status, and detail_status.
	Use process_time from time to get CPU time; perf_counter from time to get execution time.
	Execution timers are within CPU timers.
	Times are retrieved:
		before import
		after import (= Import time)
		after August setup (= Import + Setup time)
		within loop before and then after executing api calls.
		after loop to retrieve lock information (= Import + Setup + Loop)
	CPU timers retrieved before execution timer and then after execution timer

Result:
	Using session and combining door_status with lock results in 71 seconds faster execution time with 7.4 seconds less CPU time
	Using session results in 63 seconds faster execution time with 6.8 seconds less CPU time.

	Using session results in each request being 213ms faster compared to not using session.

```
Here is the test program I used:
[August_Test.py.txt](https://github.com/snjoetw/py-august/files/2507707/August_Test.py.txt)

Here are the test results.
[August_Test_results.txt](https://github.com/snjoetw/py-august/files/2507702/August_Test_results.txt)

I think that these changes will be beneficial for HASS especially when there are people with doorbell and multiple locks (with these locks then also having door sensors).
Clearly to use session would then require code changing to HASS to enable it but that is something I will do if this PR is approved and merged.

@Jefronty, I understand by using Session it would then not be possible to implement your PR (although I did use your idea for combining door/lock status; thanks!!!). However I would think that this is not really to be called as a command line but rather to be called where the status is then retrieved on a regular basis. But will leave it up to owner on how he wants to proceed. :-)