@ehendrix23 2018-10-23
Updated so that the Request.Session now has to be passed as a parameter instead of it being created within the Api class itself. If none provided then Session will not be used. Makes it so that within the Api class no need to worry about closing the session.