@snjoetw 2018-11-16
Thanks! Merge this as I think performance improvement is more beneficial