@ehendrix23 2018-11-16
@snjoetw, no worries! We all do this in our spare time and sometimes there isn't much of that left. :-)

One of the next items I would like to do is create this for async. I can just change current as is (and then breaking anyone using this without async), add async methods instead where required (not breaking anything), or a complete different fork.

I prefer the 2nd option personally (add async methods) as it won't break it for anyone else yet still keep 1 going. Otherwise it is really a complete separate fork.
Let me know your preferred approach. :-)