@snjoetw 2018-11-16
I'm actually not familiar with how async works (I'm a python newbie (-: )
What benefit does it give you if you're in an environment like HASS?