@ehendrix23 2018-11-16
Well,

Rather a newbie at Python as well but here is my understanding.

Python had multi-threading but it wasn't really multi-threading as such. The issue is that the Python interpreter itself is not multi-threaded and thus in reality only 1 thread at a time would run.
But there would still be the cost with switching threads as such within the OS.

What they also noticed was that most Python apps would be or processing something or waiting on something (i.e. input, request from a web site, ...). Thus they implemented asyncio making it easy for something to state that it is going to wait (i.e. a request) and that something else can run now.

Advantage of this is that one does not need to implement multi-threading when it really is all just about wanting to make sure that the process is not blocked just cause something has to wait on something.
And if no multi-threading then no need to lock and no cost of the multi-threading as such.

HASS has switched to asyncio and noticed a good performance improvement. Clearly they still also allow sync as not everything is changed to asyncio.

I would say that the august component is one of those that definitely would be good for asyncio.

More then willing to do it. :-)

Here is a good youtube related to asyncio:
https://www.youtube.com/watch?v=M-UcUs7IMIM