@2016for 2018-12-06
@snjoetw check out this amazing video on what exactly is Async

https://youtube.com/watch?v=tSLDcRkgTsY

Pretty nice video